More graceful handling of large files
* Use 64 bit integer for total point cloud size, and read this from the
  las extended number of points record.  The actual number of points
  loaded remains a size_t.

* Catch the std::bad_alloc which is generated when we try to allocate
  too much memory and issue an error message.  This isn't really a great
  way to handle it, but it's better than crashing.

* Beef up error reporting for files which fail to load

+++ src/pointarray.h
@@ -72,17 +72,17 @@ class PointArray : public Geometry
     private:
         bool loadLas(QString fileName, size_t maxPointCount,
                      std::vector<GeomField>& fields, V3d& offset,
-                     size_t& npoints, size_t& totPoints,
+                     size_t& npoints, uint64_t& totPoints,
                      Imath::Box3d& bbox, V3d& centroid);
 
         bool loadText(QString fileName, size_t maxPointCount,
                       std::vector<GeomField>& fields, V3d& offset,
-                      size_t& npoints, size_t& totPoints,
+                      size_t& npoints, uint64_t& totPoints,
                       Imath::Box3d& bbox, V3d& centroid);
 
         bool loadPly(QString fileName, size_t maxPointCount,
                      std::vector<GeomField>& fields, V3d& offset,
-                     size_t& npoints, size_t& totPoints,
+                     size_t& npoints, uint64_t& totPoints,
                      Imath::Box3d& bbox, V3d& centroid);
 
         friend struct ProgressFunc;
